# api-swagger.yaml
openapi: 3.0.3
info:
  title: OpenMemory API
  description: |
    OpenMemory is a memory management system with advanced features including:
    - Memory storage and retrieval
    - Temporal fact tracking
    - Reflection and insights
    - IDE integration
    - Compression utilities
    - Dynamic salience calculations
  version: 2.0.0
  contact:
    name: OpenMemory
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.openmemory.ai
    description: Production server
...
paths:
...
  /reflection/start:
    post:
      tags:
        - Reflection
      summary: Start reflection task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                time_range_hours:
                  type: integer
                  default: 24
      responses:
        '200':
          description: Reflection task started
        '400':
          description: User ID required
        '500':
          description: Internal server error
  /reflection/{user_id}/tasks:
    get:
      tags:
        - Reflection
      summary: Get user reflection tasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reflection tasks
        '400':
          description: User ID required
        '500':
          description: Internal server error
  /reflection/task/{task_id}:
    get:
      tags:
        - Reflection
      summary: Get reflection task details
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
        '400':
          description: Task ID required
        '404':
          description: Task not found
        '500':
          description: Internal server error
  /reflection/{user_id}/recent:
    get:
      tags:
        - Reflection
      summary: Get recent reflections
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Recent reflections
        '400':
          description: User ID required or invalid limit
        '500':
          description: Internal server error
  /reflection/{user_id}/search:
    post:
      tags:
        - Reflection
      summary: Search reflections by similarity
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                limit:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 100
                min_similarity:
                  type: number
                  default: 0.5
                  minimum: 0
                  maximum: 1
      responses:
        '200':
          description: Search results
        '400':
          description: User ID or query required, or invalid parameters
        '500':
          description: Internal server error
...